Player Moevemnt and Level Win
Got player movement working and got the goal platform win condition to work with options to go to the next level or back to the main menu you can see an example of it in the world 1 level 1 map. Player movement still needs simulating physics set up and the character is currently not orientated correctly because the mesh for it needs to be re-imported.