Started level design and UI work
Started designing level 1 for world 2. Started to work on UI. Have a
temporary heart placeholders with no functionality yet. Have the total
number of medicine bottles in a level and the amount collected
displaying in UI too. Created a phlegm medicine bottle blueprint that
destroys itself on collision with player