Updated Level Spawning
Updated level spawning so that the player can choose where to spawn the level. Refactored the rest of the code base to support this change. NOTE: The reposition level function in the settings menu will allow the player to reposition the level to another detected plane but it will now restart the level as well.